# Optimize mem utils functions