--- a/dbms/src/Functions/CollationStringOptimized.cpp
+++ b/dbms/src/Functions/CollationStringOptimized.cpp
@@ -22,6 +22,12 @@
 #include <cstdint>
 #include <limits>
 
+#ifdef TIFLASH_ENABLE_AVX_SUPPORT
+#ifndef TIFLASH_USE_AVX2_COMPILE_FLAG
+static_assert(false, "`dbms/src/Functions/CollationStringOptimized.cpp` need flag `-mavx2` to inline related functions");
+#endif
+#endif
+
 namespace DB
 {