### What problem does this PR solve?

Issue Number: ref #5294

### What is changed and how it works?

- add functions `avx2_strstr`, `avx2_mem_equal`, `avx2_memchr`
  - `avx2_strstr` is same as `std::string_view::find`
  - `avx2_mem_equal` is same as `bcmp` or `std::memcmp(p1,p2,n) == 0`
  - `avx2_memchr` is same as `std::memchr`
- optimize string equality comparison. https://github.com/pingcap/tiflash/blob/eaf1a4cf886d1d5a86d04ef14eff956778ab4911/libs/libcommon/include/common/StringRef.h#L93 is much slower than std::memcmp if str size is not very big.
  - according to the test result, if str-size is more than `1000000`, instructions about avx512 may begin to get better results
- optimize expression `like`

### Benchmark

ENV
- tpch-100
- 1 tiflash
- limit cpu up to 200%
- x86-64/amd64
- original commit: 8404e6565e3ff2e47e50497550d838ea76e6e776

---

https://github.com/pingcap/tiflash/blob/eef9e22ecdfd6cb8a34c353eaf5ad1dba8406981/dbms/src/Functions/tests/bench_collation.cpp#L129-L147

Time(ns) | Original | Optimized |   | Improvement: (Original) / (Optimized) - 1.0
-- | -- | -- | -- | --
CollationEqBench/UTF8MB4_BIN | 12428711 | 6228798 |   | 99.54%
CollationEqBench/UTF8_BIN | 12956705 | 6141843 |   | 110.96%
CollationEqBench/ASCII_BIN | 12625723 | 6229335 |   | 102.68%
CollationEqBench/BINARY | 11870078 | 5837615 |   | 103.34%
CollationEqBench/LATIN1_BIN | 13768201 | 6732640 |   | 104.50%
CollationLikeBench/UTF8MB4_BIN | 37940667 | 20185747 |   | 87.96%
CollationLikeBench/UTF8_BIN | 37803575 | 19914106 |   | 89.83%
CollationLikeBench/ASCII_BIN | 36860160 | 17999743 |   | 104.78%
CollationLikeBench/BINARY | 37449881 | 17599053 |   | 112.79%
CollationLikeBench/LATIN1_BIN | 37503432 | 17675036 |   | 112.18%

----

- test `bcmp`, [mem_utils::memoryEqual](https://github.com/pingcap/tiflash/blob/eaf1a4cf886d1d5a86d04ef14eff956778ab4911/libs/libcommon/include/common/mem_utils.h#L138)(use avx512) and `avx2_mem_equal`
- test `std::string_view::find` and `avx2_strstr`
- `MemUtilsEqual_xxx` means test str-size is `xxx`
- `MemUtilsStrStr_xxx_yyy` means test `src-str-size` is `xxx` and `needle-str-size` is `yyy`

Time(ns) | STL | Original-avx512 | Optimized-avx2 | Improvement: (STL) / (Optimized) - 1.0 | Improvement: (Original) / (Optimized) - 1.0
-- | -- | -- | -- | -- | --
check mem eq: MemUtilsEqual_${str-size} |   |   |   |   |  
MemUtilsEqual_13 | 4.46 | 7.22 | 4.15 | 7.47% | 73.98%
MemUtilsEqual_65 | 4.88 | 8.69 | 4.31 | 13.23% | 101.62%
MemUtilsEqual_100 | 9.9 | 13.3 | 5.65 | 75.22% | 135.40%
MemUtilsEqual_10000 | 268 | 323 | 162 | 65.43% | 99.38%
MemUtilsEqual_100000 | 3939 | 4353 | 3462 | 13.78% | 25.74%
MemUtilsEqual_1000000 | 62265 | 53157 | 52600 | 18.37% | 1.06%
  |   |   |   |   |  
str find: MemUtilsStrStr_${src-str-size}_${needle-str-size} |   |   |   |   |  
MemUtilsStrStr_1024_1 | 30882 |   | 21275 | 45.16% |  
MemUtilsStrStr_1024_7 | 34927 |   | 21279 | 64.14% |  
MemUtilsStrStr_1024_15 | 39364 |   | 23161 | 69.96% |  
MemUtilsStrStr_1024_31 | 40628 |   | 29435 | 38.03% |  
MemUtilsStrStr_1024_63 | 37381 |   | 26141 | 43.00% |  
MemUtilsStrStr_80_1 | 6130 |   | 3977 | 54.14% |  
MemUtilsStrStr_80_7 | 11720 |   | 6278 | 86.68% |  
MemUtilsStrStr_80_15 | 11585 |   | 5423 | 113.63% |  
MemUtilsStrStr_80_31 | 11467 |   | 9530 | 20.33% |  

----

SQL
```
select count(1) from orders where o_comment like '%pending%deposits%';
```

Time(s) | Original | Optimized |   |   | Improvement
-- | -- | -- | -- | -- | --
  | 10.75 | 8.72 |   |   |  
  | 10.92 | 8.87 |   |   |  
  | 10.98 | 9.05 |   |   |  
  | 10.7 | 8.64 |   |   |  
  | 10.77 | 8.78 |   |   | AVG(Original) / AVG(Optimized) - 1.0
AVG | 10.824 | 8.812 |   | Optimized : Original | 22.83%

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [x] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```