@solotzg 2022-08-25
/rebuild